drop index.js
